move intgration tests into dwa module
