+++ README.md
@@ -361,10 +361,11 @@ To benchmark the library, simply run `cargo bench`.
 
 The DWA DCT/IDCT code has SIMD kernels (AVX2 and SSE2) that are selected at
 runtime via [`pulp`](https://crates.io/crates/pulp), mirroring OpenEXR's own
-cpuid dispatch. Each tier has its own feature-gated integration test:
+cpuid dispatch. Each tier has its own feature-gated unit-test module in
+`src/compression/dwa/idct.rs`:
 
-- `tests/avx2.rs` — requires the `avx2-tests` feature, exercises the AVX2 kernel
-- `tests/sse2.rs` — requires the `sse2-tests` feature, exercises the SSE2 kernel
+- `mod avx2_tests` — requires the `avx2-tests` feature, exercises the AVX2 kernel
+- `mod sse2_tests` — requires the `sse2-tests` feature, exercises the SSE2 kernel
 
 These tests are **opt-in** and are excluded from the normal `cargo test` run. A
 test only passes if the requested tier is actually available on the CPU it runs
@@ -382,28 +383,36 @@ x86/x86-64 binary, and SDE only changes the runtime CPUID/features seen by
    binary on your `PATH` (or note its full path). SDE runs on x86-64 Linux,
    macOS, and Windows.
 
-2. **Build the test binary without running it,** then locate the executable:
+2. **Build the library test binary without running it,** then locate the executable:
 
    ```bash
    # AVX2
-   cargo test --test avx2 --features avx2-tests --no-run
+   cargo test --lib --features avx2-tests --no-run
 
    # SSE2
-   cargo test --test sse2 --features sse2-tests --no-run
+   cargo test --lib --features sse2-tests --no-run
    ```
 
-   Cargo prints the path to the compiled test binary (under `target/debug/deps/`).
+   Cargo prints the path to the compiled unit-test binary (under
+   `target/debug/deps/exr-<hash>`). Since these are now in-crate unit tests, you
+   select a tier by passing the module path as a test-name **filter** (there is
+   no per-file test binary).
 
-3. **Run that binary under SDE,** selecting a chip that exposes the target tier:
+3. **Run that binary under SDE,** selecting a chip that exposes the target tier
+   and filtering to the tier's test module:
 
    ```bash
    # AVX2 — Haswell is the first microarchitecture with AVX2 + FMA
-   sde64 -hsw -- target/debug/deps/avx2-<hash> --nocapture
+   sde64 -hsw -- target/debug/deps/exr-<hash> --nocapture compression::dwa::idct::avx2_tests
 
    # SSE2 — Merom is the lowest 64-bit chip SDE models: it has SSE2 but no AVX,
    #        so pulp falls back to the SSE2 kernel
-   sde64 -mrm -- target/debug/deps/sse2-<hash> --nocapture
+   sde64 -mrm -- target/debug/deps/exr-<hash> --nocapture compression::dwa::idct::sse2_tests
    ```
+
+   The CI workflow asserts the run reports the exact expected number of passing
+   tests, so a mistyped filter (which libtest would report as `0 passed` while
+   still exiting `0`) fails the job instead of silently passing.
 
 > [!IMPORTANT]
 > Do **not** set a global `RUSTFLAGS="-C target-feature=+avx2"` (or similar) to
@@ -414,5 +423,12 @@ x86/x86-64 binary, and SDE only changes the runtime CPUID/features seen by
 > needed. SDE alone controls which tier the runtime dispatch selects.
 
 If your own CPU already exposes the required tier, you can skip SDE and run the
-test directly (e.g. `cargo test --test sse2 --features sse2-tests` on any
-x86-64 machine, since SSE2 is part of the baseline).
+tests directly, e.g. on an AVX2-capable machine:
+
+```bash
+cargo test --lib --features avx2-tests -- compression::dwa::idct::avx2_tests
+```
+
+The SSE2 module additionally asserts that AVX2 is *absent*, so it only passes
+under SDE (or a CPU without AVX2) — running it directly on a modern machine will
+trip that assertion by design.
